+++ apps/ionic-demo/src/app/tab1/tab1.module.ts
@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { Tab1Page } from './tab1.page';
 import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
-import { NgParticlesModule } from 'ng-particles';
+import { NgxParticlesModule } from '@tsparticles/angular';
 
 import { Tab1PageRoutingModule } from './tab1-routing.module';
 
@@ -15,7 +15,7 @@ import { Tab1PageRoutingModule } from './tab1-routing.module';
     FormsModule,
     ExploreContainerComponentModule,
     Tab1PageRoutingModule,
-    NgParticlesModule
+    NgxParticlesModule
   ],
   declarations: [Tab1Page]
 })