+++ components/confetti/projects/ng-confetti/src/lib/ng-confetti.component.ts
@@ -1,14 +1,14 @@
 import { isPlatformServer } from '@angular/common';
 import { AfterViewInit, Component, Inject, Input, OnChanges, PLATFORM_ID, SimpleChanges } from '@angular/core';
 import { Subject } from 'rxjs';
-import { confetti, ConfettiOptions } from 'tsparticles-confetti';
-import type { Container } from 'tsparticles-engine';
+import { confetti, ConfettiOptions } from '@tsparticles/confetti';
+import type { Container } from '@tsparticles/engine';
 
 @Component({
-    selector: 'ng-confetti',
+    selector: 'ngx-confetti',
     template: '<div [id]="id"></div>',
 })
-export class NgConfettiComponent implements AfterViewInit, OnChanges {
+export class NgxConfettiComponent implements AfterViewInit, OnChanges {
     @Input() options?: ConfettiOptions;
     @Input() id: string;
     @Input() fire: boolean | number;