+++ components/fireworks/projects/ng-fireworks/src/lib/ng-fireworks.component.ts
@@ -1,13 +1,13 @@
 import { isPlatformServer } from '@angular/common';
 import { Component, Inject, Input, PLATFORM_ID } from '@angular/core';
-import { FireworkOptions, fireworks } from 'tsparticles-fireworks';
+import { FireworkOptions, fireworks } from '@tsparticles/fireworks';
 
 @Component({
-    selector: 'ng-fireworks',
+    selector: 'ngx-fireworks',
     template: ` <div [id]="id"></div>`,
     styles: [],
 })
-export class NgFireworksComponent {
+export class NgxFireworksComponent {
     @Input() options?: FireworkOptions;
     @Input() id: string;