+++ components/particles/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 3.0.0 (2023-12-25)
+
+
+### Bug Fixes
+
+* **deps:** update angular monorepo ([0df2410](https://github.com/tsparticles/angular/commit/0df2410d75e7a5b44766f44cf81c1d38478b33a9))
+* **deps:** update angular monorepo ([7734c62](https://github.com/tsparticles/angular/commit/7734c62fa1f0dba72675296a58b5f9e99f45c790))
+* **deps:** update angular monorepo to ~15.2.0 ([d9cc057](https://github.com/tsparticles/angular/commit/d9cc0579949f025c7be095b92abb121f2f450036))
+* **deps:** update angular monorepo to ~16.2.0 ([c94af32](https://github.com/tsparticles/angular/commit/c94af3266705e56cdd68fd9709882f97bf0a95a1))
+* **deps:** update angular monorepo to v16 ([ca8f4e4](https://github.com/tsparticles/angular/commit/ca8f4e47c4476a5f4548ad25f365f56caa7ce912))
+* **deps:** update dependency zone.js to ~0.13.0 ([9eaf8c1](https://github.com/tsparticles/angular/commit/9eaf8c1e184fbe813e0f2ec4ed14d35c677bc8a3))
+
+
+### Features
+
+* added service for initializing particles plugins ([c26e9f2](https://github.com/tsparticles/angular/commit/c26e9f2caa2038e652f2967a38ac26c6b5264a49))
+* preparing confetti and fireworks components ([a098d21](https://github.com/tsparticles/angular/commit/a098d217f7f332bfef5bda4d6d846b762ee4aea3))
+
+
+
+
+
 # [3.12.0](https://github.com/tsparticles/angular/compare/ng-particles@3.11.0...ng-particles@3.12.0) (2023-08-04)
 
 **Note:** Version bump only for package ng-particles
@@ -426,7 +448,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 -   fixed angular readme, closes [#2297](https://github.com/matteobruni/tsparticles/issues/2297) ([5f51b87](https://github.com/matteobruni/tsparticles/commit/5f51b87a6fca58ee88ae51a17dbd5e0d1911b3c6))
 -   fixed some readmes ([93f371a](https://github.com/matteobruni/tsparticles/commit/93f371ab82a5074d34ec7632ade41edc3dbf0ec7))
 
-# [3.0.0-beta.2](https://github.com/matteobruni/tsparticles/compare/ng-particles@2.18.4...ng-particles@3.0.0-beta.2) (2021-10-06)
+# [3.0.0-beta.3](https://github.com/matteobruni/tsparticles/compare/ng-particles@2.18.4...ng-particles@3.0.0-beta.3) (2021-10-06)
 
 ## [2.20.5](https://github.com/matteobruni/tsparticles/compare/ng-particles@2.20.4...ng-particles@2.20.5) (2021-11-28)